refactor operator admin routes
